+++ R/wfs-svm-poly-reg.R
@@ -44,24 +44,11 @@
 #' suppressPackageStartupMessages(library(modeltime))
 #' suppressPackageStartupMessages(library(timetk))
 #' suppressPackageStartupMessages(library(dplyr))
-#' suppressPackageStartupMessages(library(healthyR.data))
 #' suppressPackageStartupMessages(library(tidymodels))
-#' suppressPackageStartupMessages(library(earth))
 #'
-#' data <- healthyR_data %>%
-#'  filter(ip_op_flag == "I") %>%
-#'    select(visit_end_date_time) %>%
-#'    rename(date_col = visit_end_date_time) %>%
-#'    summarise_by_time(
-#'        .date_var = date_col
-#'        , .by     = "month"
-#'        , value   = n()
-#'   ) %>%
-#'    filter_by_time(
-#'        .date_var     = date_col
-#'        , .start_date = "2012"
-#'        , .end_date   = "2019"
-#'    )
+#' data <- AirPassengers %>%
+#'   ts_to_tbl() %>%
+#'   select(-index)
 #'
 #' splits <- time_series_split(
 #'    data